Resolved 8 conflicts:
- README.md: take origin/main (eFuse warning wording)
- components/cryptnox_utils/CMakeLists.txt: take origin/main (esp_log→log, bt as optional)
- components/cryptnox_utils/esp32_random.cpp: take origin/main (BT guard, LOGW, always fill)
- components/crypto_provider/esp32_crypto_provider.cpp: take origin/main (ecdsaVerify with curve)
- components/crypto_provider/include/esp32_crypto_provider.h: take origin/main (ecdsaVerify sig)
- components/secure_channel/CMakeLists.txt: keep HEAD (CW_DEBUG_LOGGING commented out, SEC-012)
- components/secure_channel/test/test_cw_secure_channel.cpp: take origin/main (remove uECC.h)
- sdkconfig.defaults: take origin/main (security configs + stack size 8192)

+++ components/pn532/include/pn532.h
@@ -34,7 +34,7 @@ uint32_t  pn532_get_firmware_version(pn532_t *dev);
 bool      pn532_sam_config(pn532_t *dev);
 uint32_t  pn532_read_passive_target_id(pn532_t *dev, uint8_t cardbaudrate);
 bool      pn532_send_apdu(pn532_t *dev, const uint8_t *apdu, uint8_t apdu_len,
-                          uint8_t *response, uint8_t *response_len);
+                          uint8_t *response, uint16_t *response_len);
 bool      pn532_release_target(pn532_t *dev);
 
 #ifdef __cplusplus
